requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy>=2.0
python-multipart>=0.0.18
jinja2>=3.1
python-docx>=1.1
PyPDF2>=3.0
google-genai>=1.14
httpx>=0.28
python-dotenv>=1.0
